Wrote weather dashboard as basically as possible.
Didn't use any AI intervention until I got to generating weather cards and plotting historical data.

I did read all of the relevant api documentation and wrote my own api calls to retrieve the data being used. I used javascript to pass in variables like latitude and longitude.

I did use AI for styling this thing. That's an area I couold've spent a ton of time and so I figured I would see what AI could do. This was specifically ChatGPT 5.

Chat GPT did write the javascript for generating the forecast cards like I previously stated. I didn't really mean for it to take the reigns but it kinda just wrote it all without me asking for it. You'll have to take my word for it when I say I understand what it is doing and how it is generating the forecast cards.

This was a tough project with a lot of moving parts. I probably could've dedicated a couple day's to coding this if not for using AI. Though I do believe the javascript it wrote is overly complex and this could've been achieved with less code.